+++ ic-agent/src/agent/agent_config.rs
@@ -1,11 +1,10 @@
 use crate::{
-    agent::{NonceFactory, NonceGenerator, ReplicaV2Transport},
+    agent::{NonceFactory, NonceGenerator, Transport},
     identity::{anonymous::AnonymousIdentity, Identity},
 };
 use std::{sync::Arc, time::Duration};
 
 /// A configuration for an agent.
-#[derive(Debug)]
 pub struct AgentConfig {
     /// See [`with_nonce_factory`](super::AgentBuilder::with_nonce_factory).
     pub nonce_factory: Arc<dyn NonceGenerator>,
@@ -14,7 +13,7 @@ pub struct AgentConfig {
     /// See [`with_ingress_expiry`](super::AgentBuilder::with_ingress_expiry).
     pub ingress_expiry_duration: Option<Duration>,
     /// The [`with_transport`](super::AgentBuilder::with_transport).
-    pub transport: Option<Arc<dyn ReplicaV2Transport>>,
+    pub transport: Option<Arc<dyn Transport>>,
 }
 
 impl Default for AgentConfig {