+++ ic-agent/src/agent/agent_test.rs
@@ -1,129 +1,126 @@
 // Disable these tests without the reqwest feature.
 #![cfg(feature = "reqwest")]
 
+use self::mock::{assert_mock, mock};
 use crate::{
     agent::{
-        http_transport::ReqwestHttpReplicaV2Transport,
-        replica_api::{CallReply, QueryResponse},
+        http_transport::ReqwestTransport,
+        replica_api::{CallReply, QueryResponse, RejectCode, RejectResponse},
         Status,
     },
     export::Principal,
     Agent, AgentError,
 };
 use ic_certification::Label;
-use mockito::mock;
 use std::collections::BTreeMap;
+#[cfg(target_family = "wasm")]
+use wasm_bindgen_test::wasm_bindgen_test;
 
-#[test]
-fn query() -> Result<(), AgentError> {
+#[cfg(target_family = "wasm")]
+wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_browser);
+
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
+async fn query() -> Result<(), AgentError> {
     let blob = Vec::from("Hello World");
     let response = QueryResponse::Replied {
         reply: CallReply { arg: blob.clone() },
     };
 
-    let query_mock = mock("POST", "/api/v2/canister/aaaaa-aa/query")
-        .with_status(200)
-        .with_header("content-type", "application/cbor")
-        .with_body(serde_cbor::to_vec(&response)?)
-        .create();
+    let (query_mock, url) = mock(
+        "POST",
+        "/api/v2/canister/aaaaa-aa/query",
+        200,
+        serde_cbor::to_vec(&response)?,
+        Some("application/cbor"),
+    )
+    .await;
 
     let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(
-            &mockito::server_url(),
-        )?)
+        .with_transport(ReqwestTransport::create(&url)?)
         .build()?;
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-    let result = runtime.block_on(async {
-        agent
-            .query_raw(
-                &Principal::management_canister(),
-                Principal::management_canister(),
-                "main",
-                &[],
-                None,
-            )
-            .await
-    });
+    let result = agent
+        .query_raw(
+            &Principal::management_canister(),
+            Principal::management_canister(),
+            "main",
+            &[],
+            None,
+        )
+        .await;
 
-    query_mock.assert();
+    assert_mock(query_mock).await;
 
     assert_eq!(result?, blob);
 
     Ok(())
 }
 
-#[test]
-fn query_error() -> Result<(), AgentError> {
-    let query_mock = mock("POST", "/api/v2/canister/aaaaa-aa/query")
-        .with_status(500)
-        .create();
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
+async fn query_error() -> Result<(), AgentError> {
+    let (query_mock, url) =
+        mock("POST", "/api/v2/canister/aaaaa-aa/query", 500, vec![], None).await;
     let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(
-            &mockito::server_url(),
-        )?)
+        .with_transport(ReqwestTransport::create(url)?)
         .build()?;
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
 
-    let result = runtime.block_on(async {
-        agent
-            .query_raw(
-                &Principal::management_canister(),
-                Principal::management_canister(),
-                "greet",
-                &[],
-                None,
-            )
-            .await
-    });
+    let result = agent
+        .query_raw(
+            &Principal::management_canister(),
+            Principal::management_canister(),
+            "greet",
+            &[],
+            None,
+        )
+        .await;
 
-    query_mock.assert();
+    assert_mock(query_mock).await;
 
     assert!(result.is_err());
 
     Ok(())
 }
 
-#[test]
-fn query_rejected() -> Result<(), AgentError> {
-    let response: QueryResponse = QueryResponse::Rejected {
-        reject_code: 1234,
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
+async fn query_rejected() -> Result<(), AgentError> {
+    let response: QueryResponse = QueryResponse::Rejected(RejectResponse {
+        reject_code: RejectCode::DestinationInvalid,
         reject_message: "Rejected Message".to_string(),
-    };
-
-    let query_mock = mock("POST", "/api/v2/canister/aaaaa-aa/query")
-        .with_status(200)
-        .with_header("content-type", "application/cbor")
-        .with_body(serde_cbor::to_vec(&response)?)
-        .create();
-
-    let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(
-            &mockito::server_url(),
-        )?)
-        .build()?;
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-
-    let result = runtime.block_on(async {
-        agent
-            .query_raw(
-                &Principal::management_canister(),
-                Principal::management_canister(),
-                "greet",
-                &[],
-                None,
-            )
-            .await
+        error_code: Some("Error code".to_string()),
     });
 
-    query_mock.assert();
+    let (query_mock, url) = mock(
+        "POST",
+        "/api/v2/canister/aaaaa-aa/query",
+        200,
+        serde_cbor::to_vec(&response)?,
+        Some("application/cbor"),
+    )
+    .await;
+
+    let agent = Agent::builder()
+        .with_transport(ReqwestTransport::create(&url)?)
+        .build()?;
+
+    let result = agent
+        .query_raw(
+            &Principal::management_canister(),
+            Principal::management_canister(),
+            "greet",
+            &[],
+            None,
+        )
+        .await;
+
+    assert_mock(query_mock).await;
 
     match result {
-        Err(AgentError::ReplicaError {
-            reject_code: code,
-            reject_message: msg,
-        }) => {
-            assert_eq!(code, 1234);
-            assert_eq!(msg, "Rejected Message");
+        Err(AgentError::ReplicaError(replica_error)) => {
+            assert_eq!(replica_error.reject_code, RejectCode::DestinationInvalid);
+            assert_eq!(replica_error.reject_message, "Rejected Message");
+            assert_eq!(replica_error.error_code, Some("Error code".to_string()));
         }
         result => unreachable!("{:?}", result),
     }
@@ -131,36 +128,107 @@ fn query_rejected() -> Result<(), AgentError> {
     Ok(())
 }
 
-#[test]
-fn call_error() -> Result<(), AgentError> {
-    let call_mock = mock("POST", "/api/v2/canister/aaaaa-aa/call")
-        .with_status(500)
-        .create();
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
+async fn call_error() -> Result<(), AgentError> {
+    let (call_mock, url) = mock("POST", "/api/v2/canister/aaaaa-aa/call", 500, vec![], None).await;
 
     let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(
-            &mockito::server_url(),
-        )?)
+        .with_transport(ReqwestTransport::create(&url)?)
         .build()?;
 
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-    let result = runtime.block_on(async {
-        agent
-            .update(&Principal::management_canister(), "greet")
-            .with_arg([])
-            .call()
-            .await
-    });
+    let result = agent
+        .update(&Principal::management_canister(), "greet")
+        .with_arg([])
+        .call()
+        .await;
 
-    call_mock.assert();
+    assert_mock(call_mock).await;
 
     assert!(result.is_err());
 
     Ok(())
 }
 
-#[test]
-fn status() -> Result<(), AgentError> {
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
+async fn call_rejected() -> Result<(), AgentError> {
+    let reject_body = RejectResponse {
+        reject_code: RejectCode::SysTransient,
+        reject_message: "Test reject message".to_string(),
+        error_code: Some("Test error code".to_string()),
+    };
+
+    let body = serde_cbor::to_vec(&reject_body).unwrap();
+
+    let (call_mock, url) = mock(
+        "POST",
+        "/api/v2/canister/aaaaa-aa/call",
+        200,
+        body,
+        Some("application/cbor"),
+    )
+    .await;
+
+    let agent = Agent::builder()
+        .with_transport(ReqwestTransport::create(&url)?)
+        .build()?;
+
+    let result = agent
+        .update(&Principal::management_canister(), "greet")
+        .with_arg([])
+        .call()
+        .await;
+
+    assert_mock(call_mock).await;
+
+    let expected_response = Err(AgentError::ReplicaError(reject_body));
+    assert_eq!(expected_response, result);
+
+    Ok(())
+}
+
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
+async fn call_rejected_without_error_code() -> Result<(), AgentError> {
+    let reject_body = RejectResponse {
+        reject_code: RejectCode::SysTransient,
+        reject_message: "Test reject message".to_string(),
+        error_code: None,
+    };
+
+    let body = serde_cbor::to_vec(&reject_body).unwrap();
+
+    let (call_mock, url) = mock(
+        "POST",
+        "/api/v2/canister/aaaaa-aa/call",
+        200,
+        body,
+        Some("application/cbor"),
+    )
+    .await;
+
+    let agent = Agent::builder()
+        .with_transport(ReqwestTransport::create(&url)?)
+        .build()?;
+
+    let result = agent
+        .update(&Principal::management_canister(), "greet")
+        .with_arg([])
+        .call()
+        .await;
+
+    assert_mock(call_mock).await;
+
+    let expected_response = Err(AgentError::ReplicaError(reject_body));
+    assert_eq!(expected_response, result);
+
+    Ok(())
+}
+
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
+async fn status() -> Result<(), AgentError> {
     let ic_api_version = "1.2.3".to_string();
     let mut map = BTreeMap::new();
     map.insert(
@@ -168,70 +236,71 @@ fn status() -> Result<(), AgentError> {
         serde_cbor::Value::Text(ic_api_version.clone()),
     );
     let response = serde_cbor::Value::Map(map);
-    let read_mock = mock("GET", "/api/v2/status")
-        .with_status(200)
-        .with_body(serde_cbor::to_vec(&response)?)
-        .create();
+    let (read_mock, url) = mock(
+        "GET",
+        "/api/v2/status",
+        200,
+        serde_cbor::to_vec(&response)?,
+        Some("application/cbor"),
+    )
+    .await;
 
     let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(
-            &mockito::server_url(),
-        )?)
+        .with_transport(ReqwestTransport::create(&url)?)
         .build()?;
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-    let result = runtime.block_on(async { agent.status().await });
+    let result = agent.status().await;
 
-    read_mock.assert();
+    assert_mock(read_mock).await;
     assert!(matches!(result, Ok(Status { ic_api_version: v, .. }) if v == ic_api_version));
 
     Ok(())
 }
 
-#[test]
-fn status_okay() -> Result<(), AgentError> {
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
+async fn status_okay() -> Result<(), AgentError> {
     let mut map = BTreeMap::new();
     map.insert(
         serde_cbor::Value::Text("ic_api_version".to_owned()),
         serde_cbor::Value::Text("1.2.3".to_owned()),
     );
     let response = serde_cbor::Value::Map(map);
-    let read_mock = mock("GET", "/api/v2/status")
-        .with_status(200)
-        .with_body(serde_cbor::to_vec(&response)?)
-        .create();
+    let (read_mock, url) = mock(
+        "GET",
+        "/api/v2/status",
+        200,
+        serde_cbor::to_vec(&response)?,
+        Some("application/cbor"),
+    )
+    .await;
 
     let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(
-            &mockito::server_url(),
-        )?)
+        .with_transport(ReqwestTransport::create(&url)?)
         .build()?;
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-    let result = runtime.block_on(agent.status());
+    let result = agent.status().await;
 
-    read_mock.assert();
+    assert_mock(read_mock).await;
 
     assert!(result.is_ok());
 
     Ok(())
 }
 
-#[test]
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
 // test that the agent (re)tries to reach the server.
 // We spawn an agent that waits 400ms between requests, and times out after 600ms. The agent is
 // expected to hit the server at ~ 0ms and ~ 400 ms, and then shut down at 600ms, so we check that
 // the server got two requests.
-fn status_error() -> Result<(), AgentError> {
+async fn status_error() -> Result<(), AgentError> {
     // This mock is never asserted as we don't know (nor do we need to know) how many times
     // it is called.
-    let _read_mock = mock("GET", "/api/v2/status").with_status(500).create();
+    let (_read_mock, url) = mock("GET", "/api/v2/status", 500, vec![], None).await;
 
     let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(
-            &mockito::server_url(),
-        )?)
+        .with_transport(ReqwestTransport::create(&url)?)
         .build()?;
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-    let result = runtime.block_on(async { agent.status().await });
+    let result = agent.status().await;
 
     assert!(result.is_err());
 
@@ -357,98 +426,206 @@ const PRUNED_SUBNET: [u8; 1064] = [
     178, 247, 106, 100, 101, 108, 101, 103, 97, 116, 105, 111, 110, 246,
 ];
 
-#[test]
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
 // asserts that a delegated certificate with correct /subnet/<subnetid>/canister_ranges
 // passes the certificate verification
-fn check_subnet_range_with_valid_range() {
-    let _read_mock = mock(
+async fn check_subnet_range_with_valid_range() {
+    let (_read_mock, url) = mock(
         "POST",
         "/api/v2/canister/ivg37-qiaaa-aaaab-aaaga-cai/read_state",
+        200,
+        REQ_WITH_DELEGATED_CERT_RESPONSE.into(),
+        Some("application/cbor"),
     )
-    .with_status(200)
-    .with_body(REQ_WITH_DELEGATED_CERT_RESPONSE)
-    .create();
+    .await;
     let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(&mockito::server_url()).unwrap())
+        .with_transport(ReqwestTransport::create(&url).unwrap())
         .build()
         .unwrap();
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-    let _result = runtime
-        .block_on(async {
-            agent
-                .read_state_raw(
-                    vec![REQ_WITH_DELEGATED_CERT_PATH
-                        .iter()
-                        .map(Label::from)
-                        .collect()],
-                    Principal::from_text(REQ_WITH_DELEGATED_CERT_CANISTER).unwrap(),
-                )
-                .await
-        })
+    let _result = agent
+        .read_state_raw(
+            vec![REQ_WITH_DELEGATED_CERT_PATH
+                .iter()
+                .map(Label::from)
+                .collect()],
+            Principal::from_text(REQ_WITH_DELEGATED_CERT_CANISTER).unwrap(),
+        )
+        .await
         .expect("read state failed");
 }
 
-#[test]
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
 // asserts that a delegated certificate with /subnet/<subnetid>/canister_ranges that don't include
 // the canister gets rejected by the cert verification because the subnet is not authorized to
 // respond to requests for this canister. We do this by using a correct response but serving it
 // for the wrong canister, which a malicious node might do.
-fn check_subnet_range_with_unauthorized_range() {
+async fn check_subnet_range_with_unauthorized_range() {
     let wrong_canister = Principal::from_text("ryjl3-tyaaa-aaaaa-aaaba-cai").unwrap();
-    let _read_mock = mock(
+    let (_read_mock, url) = mock(
         "POST",
         "/api/v2/canister/ryjl3-tyaaa-aaaaa-aaaba-cai/read_state",
+        200,
+        REQ_WITH_DELEGATED_CERT_RESPONSE.into(),
+        Some("application/cbor"),
     )
-    .with_status(200)
-    .with_body(REQ_WITH_DELEGATED_CERT_RESPONSE)
-    .create();
+    .await;
     let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(&mockito::server_url()).unwrap())
+        .with_transport(ReqwestTransport::create(&url).unwrap())
         .build()
         .unwrap();
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-    let result = runtime.block_on(async {
-        agent
-            .read_state_raw(
-                vec![REQ_WITH_DELEGATED_CERT_PATH
-                    .iter()
-                    .map(Label::from)
-                    .collect()],
-                wrong_canister,
-            )
-            .await
-    });
+    let result = agent
+        .read_state_raw(
+            vec![REQ_WITH_DELEGATED_CERT_PATH
+                .iter()
+                .map(Label::from)
+                .collect()],
+            wrong_canister,
+        )
+        .await;
     assert_eq!(result, Err(AgentError::CertificateNotAuthorized()));
 }
 
-#[test]
+#[cfg_attr(not(target_family = "wasm"), tokio::test)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen_test)]
 // asserts that a delegated certificate with pruned/removed /subnet/<subnetid>/canister_ranges
 // gets rejected by the cert verification. We do this by using a correct response that has
 // the leaf manually pruned
-fn check_subnet_range_with_pruned_range() {
+async fn check_subnet_range_with_pruned_range() {
     let canister = Principal::from_text("ivg37-qiaaa-aaaab-aaaga-cai").unwrap();
-    let _read_mock = mock(
+    let (_read_mock, url) = mock(
         "POST",
         "/api/v2/canister/ivg37-qiaaa-aaaab-aaaga-cai/read_state",
+        200,
+        PRUNED_SUBNET.into(),
+        Some("application/cbor"),
     )
-    .with_status(200)
-    .with_body(PRUNED_SUBNET)
-    .create();
+    .await;
     let agent = Agent::builder()
-        .with_transport(ReqwestHttpReplicaV2Transport::create(&mockito::server_url()).unwrap())
+        .with_transport(ReqwestTransport::create(&url).unwrap())
         .build()
         .unwrap();
-    let runtime = tokio::runtime::Runtime::new().expect("Unable to create a runtime");
-    let result = runtime.block_on(async {
-        agent
-            .read_state_raw(
-                vec![REQ_WITH_DELEGATED_CERT_PATH
-                    .iter()
-                    .map(Label::from)
-                    .collect()],
-                canister,
-            )
-            .await
-    });
+    let result = agent
+        .read_state_raw(
+            vec![REQ_WITH_DELEGATED_CERT_PATH
+                .iter()
+                .map(Label::from)
+                .collect()],
+            canister,
+        )
+        .await;
     assert!(result.is_err());
+}
+
+#[cfg(not(target_family = "wasm"))]
+mod mock {
+
+    use mockito::{Mock, Server, ServerGuard};
+
+    pub async fn mock(
+        method: &str,
+        path: &str,
+        status_code: u16,
+        body: Vec<u8>,
+        content_type: Option<&str>,
+    ) -> ((ServerGuard, Mock), String) {
+        let mut server = Server::new_async().await;
+        let mut mock = server
+            .mock(method, path)
+            .with_status(status_code as _)
+            .with_body(body);
+        if let Some(content_type) = content_type {
+            mock = mock.with_header("Content-Type", content_type);
+        }
+        let mock = mock.create_async().await;
+        let url = server.url();
+        ((server, mock), url)
+    }
+
+    pub async fn assert_mock((_, mock): (ServerGuard, Mock)) {
+        mock.assert_async().await;
+    }
+}
+
+#[cfg(target_family = "wasm")]
+mod mock {
+    use js_sys::*;
+    use reqwest::Client;
+    use serde::Serialize;
+    use std::collections::HashMap;
+    use wasm_bindgen::{prelude::*, JsCast};
+    use wasm_bindgen_futures::JsFuture;
+    use web_sys::*;
+
+    #[wasm_bindgen(module = "/http_mock_service_worker.js")]
+    extern "C" {}
+
+    #[derive(Debug, Serialize)]
+    struct MockConfig {
+        pub kind: String,
+        pub method: String,
+        pub path: String,
+        pub nonce: String,
+        pub status_code: u16,
+        pub headers: Option<HashMap<String, String>>,
+        pub body: Vec<u8>,
+    }
+
+    pub async fn mock(
+        method: &str,
+        path: &str,
+        status_code: u16,
+        body: Vec<u8>,
+        content_type: Option<&str>,
+    ) -> (String, String) {
+        let swc = window().unwrap().navigator().service_worker();
+        let registration: ServiceWorkerRegistration =
+            JsFuture::from(swc.register("/http_mock_service_worker.js"))
+                .await
+                .unwrap()
+                .unchecked_into();
+        JsFuture::from(swc.ready().unwrap()).await.unwrap();
+        let sw = registration.active().unwrap();
+        let mut nonce = [0; 16];
+        getrandom::getrandom(&mut nonce).unwrap();
+        let nonce = hex::encode(nonce);
+        let config = MockConfig {
+            kind: "config".into(),
+            nonce: nonce.clone(),
+            method: method.into(),
+            path: path.into(),
+            status_code,
+            body,
+            headers: content_type.map(|c| HashMap::from([("Content-Type".into(), c.into())])),
+        };
+        if sw.state() == ServiceWorkerState::Activating {
+            JsFuture::from(Promise::new(&mut |rs, _| sw.set_onstatechange(Some(&rs))))
+                .await
+                .unwrap();
+        }
+        Client::new()
+            .post("http://mock_configure")
+            .json(&config)
+            .send()
+            .await
+            .unwrap()
+            .error_for_status()
+            .unwrap();
+        (nonce.clone(), format!("http://mock_{}/", nonce))
+    }
+
+    pub async fn assert_mock(nonce: String) {
+        let hits = Client::new()
+            .get(&format!("http://mock_assert/{}", nonce))
+            .send()
+            .await
+            .unwrap()
+            .error_for_status()
+            .unwrap()
+            .text()
+            .await
+            .unwrap();
+        assert!(hits.parse::<i32>().unwrap() >= 1);
+    }
 }