+++ ref-tests/src/utils.rs
@@ -1,4 +1,4 @@
-use ic_agent::agent::http_transport::ReqwestHttpReplicaV2Transport;
+use ic_agent::agent::http_transport::ReqwestTransport;
 use ic_agent::identity::Secp256k1Identity;
 use ic_agent::{export::Principal, identity::BasicIdentity, Agent, Identity};
 use ic_identity_hsm::HardwareIdentity;
@@ -85,9 +85,7 @@ pub async fn create_agent(identity: Box<dyn Identity>) -> Result<Agent, String> 
         .expect("Could not parse the IC_REF_PORT environment variable as an integer.");
 
     Agent::builder()
-        .with_transport(
-            ReqwestHttpReplicaV2Transport::create(format!("http://127.0.0.1:{}", port)).unwrap(),
-        )
+        .with_transport(ReqwestTransport::create(format!("http://127.0.0.1:{}", port)).unwrap())
         .with_boxed_identity(identity)
         .build()
         .map_err(|e| format!("{:?}", e))